requirements.txt
asn1crypto==1.5.1
aws-saml-login==1.0.11
beautifulsoup4==4.11.1
boto3==1.24.15
botocore==1.27.15
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==36.0.2
detect-delimiter==0.1.1
distlib==0.3.5
filelock==3.8.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
httplib2==0.20.4
idna==3.3
jmespath==1.0.1
ndg-httpsclient==0.5.1
numpy==1.23.0
oauthlib==3.2.0
oscrypto==1.3.0
pandas==1.4.3
platformdirs==2.5.2
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.15.0
pygsheets==2.0.5
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.6.0
six==1.16.0
snowflake==0.0.3
snowflake-connector-python==2.7.9
soupsieve==2.3.2.post1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.16.3
